move playback hooks to dedicated folder
