+++ app/packages/playback/src/lib/hooks/use-default-timeline-name.ts
@@ -1,7 +1,7 @@
 import * as fos from "@fiftyone/state";
 import { useCallback, useMemo } from "react";
 import { useRecoilValue } from "recoil";
-import { GLOBAL_TIMELINE_ID } from "./constants";
+import { GLOBAL_TIMELINE_ID } from "../constants";
 
 export const getTimelineNameFromSampleAndGroupId = (
   sampleId?: string | null,
